Fixing login controllers, implementing Tims merge conflicts
