+++ RevaturePetInsurance/src/main/java/com/controller/LoginController.java
@@ -4,6 +4,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -19,21 +20,37 @@ import com.service.OwnerService;
 @CrossOrigin(origins="*")
 public class LoginController {
 	
-	private OwnerDao od;
 	private OwnerService os;
 	
+	public LoginController() {
+		
+	}
+	
+	@Autowired
+	public LoginController(OwnerService os) {
+		this.os = os;
+	}
+	
 	@PostMapping(value="/login.app")
-	public @ResponseBody Owner login(HttpServletRequest req, HttpServletResponse res) {
+	public  @ResponseBody String login(HttpServletRequest req, HttpServletResponse res) {
 		
 		HttpSession ses = req.getSession();
 		String email = req.getParameter("email");
 		String password = req.getParameter("password");
+		Owner owner = os.verifyLogin(email, password);
 		
-		if(!(os.verifyLogin(email, password))) {
-			return null;
-		} else {
-			return od.selectOwnerByEmail(email);
-		}
+		if(owner.equals(null)) {
+			return "false";
+			} else {
+			ses.setAttribute("loggedUser", owner);
+			return "true";		
+	}
 	
+	}
+	
+	public void logout(HttpServletRequest req) {
+		HttpSession ses = req.getSession();
+		Owner owner = (Owner)ses.getAttribute("loggedUser");
+		ses.invalidate();
 	}
 }
