--- RevaturePetInsurance/src/main/java/com/controller/OwnerControllerTemp.java
@@ -1,30 +1,0 @@
-package com.controller;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-import com.model.Owner;
-import com.service.OwnerServiceImpl;
-
-public class OwnerControllerTemp {
-	
-	private OwnerServiceImpl os;
-	
-	public OwnerControllerTemp() {
-	}
-	
-	public OwnerControllerTemp(OwnerServiceImpl os) {
-		this.os = os;
-	}
-	
-	@PostMapping(value="/getOwner.app")
-	public @ResponseBody Owner displayOwner(HttpServletRequest req, HttpServletResponse res) {
-		HttpSession ses = req.getSession();
-		Owner owner = (Owner)ses.getAttribute("loggedUser");
-		return owner;
-	}
-}
