+++ RevaturePetInsurance/src/main/webapp/WEB-INF/applicationContext.xml
@@ -9,10 +9,19 @@
 		http://www.springframework.org/schema/mvc http://www.springframework.org/schema/mvc/spring-mvc.xsd
 		http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx.xsd">
 
+	<context:annotation-config></context:annotation-config>
+	<context:component-scan base-package="com.example"></context:component-scan>
+	
+	<!-- turns on the annotations for transactions -->
+	<tx:annotation-driven/>
+	
+	<!-- mvc annotations for transactions -->
+	<mvc:annotation-driven></mvc:annotation-driven>
+
 <bean id="dataSource"
 		class="org.springframework.jdbc.datasource.DriverManagerDataSource">
 		<property name="driverClassName"
-			value="oracle.jdbc.driver.OracleDriver" />
+			value="org.postgresql.Driver" />
 		<property name="url"
 			value="#{systemEnvironment['PetInsuranceURL']}" />
 		<property name="username"
@@ -24,7 +33,7 @@
 	<bean id="sessionFactory"
 		class="org.springframework.orm.hibernate5.LocalSessionFactoryBean">
 		<property name="dataSource" ref="dataSource" />
-		<property name="packagesToScan" value="com.example.model" />
+		<property name="packagesToScan" value="com" />
 		<property name="hibernateProperties">
 			<props>
 				<prop key="hibernate.dialect">org.hibernate.dialect.Oracle10gDialect</prop>
