+++ RevaturePetInsurance/src/main/webapp/WEB-INF/web.xml
@@ -1,16 +1,31 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://java.sun.com/xml/ns/javaee" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd" version="2.5">
-  <display-name>RevaturePetInsurance</display-name>
-  <welcome-file-list>
+<web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xmlns="http://java.sun.com/xml/ns/javaee"
+	xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd"
+	version="2.5">
+	<display-name>RevaturePetInsurance</display-name>
+	<welcome-file-list>
 		<welcome-file>index.html</welcome-file>
 	</welcome-file-list>
+
+	<servlet>
+		<!-- Whatever you put in the "servlet-name" tag, you'll need a file called 
+			"chosenname-servlet.xml" to match -->
+		<servlet-name>dispatcher</servlet-name>
+		<servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
+	</servlet>
+	<servlet-mapping>
+		<servlet-name>dispatcher</servlet-name>
+		<url-pattern>*app</url-pattern>
+	</servlet-mapping>
+
 
 	<servlet>
 		<display-name>HibernateServlet</display-name>
 		<servlet-name>HibernateServlet</servlet-name>
 		<servlet-class>com.servlet.HibernateServlet</servlet-class>
 	</servlet>
-	
+
 	<servlet-mapping>
 		<servlet-name>HibernateServlet</servlet-name>
 		<url-pattern>/Hibernate/*</url-pattern>
